requirements.txt
numpy==1.21.0
pandas==1.3.0
torch==1.9.0
transformers==4.11.3
nltk==3.6.2
pyyaml==5.4.1
pytest==6.2.5
flask==2.0.1